### 2024-05-30 #6808 @editorialbot
[PRE REVIEW]: wmpy-power: A Python package for process-based regional hydropower simulation
**Submitting author:** @thurber (<a href="http://orcid.org/0000-0002-4370-9971">Travis Thurber</a>)
**Repository:** https://github.com/9505-PNNL/wmpy_power
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
