+++ paper/paper.bib
@@ -2,7 +2,7 @@
     journal = {{in prep}},
     year = {2024},
     title = {{Multi-scale impacts of climate change on hydropower for long-term water-energy planning in the contiguous United States}},
-    author = {Broman, Daniel and  Kao, {Shih-Chieh} and Fernandez, Alisha and Ghimire, Ganesh and Voisin, Nathalie},
+    author = {Broman, Daniel and Voisin, Nathalie and  Kao, {Shih-Chieh} and Fernandez, Alisha and Ghimire, Ganesh},
 }
 
 @article{Duan:1993,
@@ -62,10 +62,10 @@
     title = {Water storage and release policies for all large reservoirs of conterminous United States}, 
     journal = {Journal of Hydrology}, 
     volume = {603}, 
-    pages = {126843}, 
+    pages = {126843},
     year = {2021}, 
     issn = {0022-1694}, 
-    doi = {https://doi.org/10.1016/j.jhydrol.2021.126843}, 
+    doi = {10.1016/j.jhydrol.2021.126843}, 
     url = {https://www.sciencedirect.com/science/article/pii/S0022169421008933}, 
     author = {Sean W.D. Turner and Jennie Clarice Steyaert and Laura Condon and Nathalie Voisin},
 }
