+++ wmpy_power/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 from .model import Model
