### 2024-07-13 #6808 @kyleniemeyer
Hi @thurber, per @AdamRJensen's recommendation, we're going to take another look at the scope of the software. I would recommend thinking about documentation - I see right now that the README is quite long with some of that, but in general it would be best to create dedicated documentation and just include some basic information in the README.
