### 2024-09-14 #6808 @ijbd
Hi,

I would like to help review this submission. I can start this coming week
if that is alright?

On Tue, Sep 10, 2024 at 11:58 AM Taher Chegini ***@***.***>
wrote:

> 👋🏼 @antviro <https://github.com/antviro> and @ijbd
> <https://github.com/ijbd>, Would you like to review this submission to
> the Journal for Open Source Software? Our reviews are checklist-driven and
> openly conducted on GitHub over a timeline of 4–6 weeks. Because the
> process is much more iterative and interactive than a traditional paper
> review, we would ask you to start within the next 2 weeks. Here are
> reviewer guidelines for reference:
> joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Thanks for your consideration.
>
