### 2024-09-14 #7225 @editorialbot
[REVIEW]: wmpy-power: A Python package for process-based regional hydropower simulation
**Submitting author:** @thurber (<a href="http://orcid.org/0000-0002-4370-9971">Travis Thurber</a>)
**Repository:** https://github.com/9505-PNNL/wmpy_power
**Branch with paper.md** (empty if default branch):
**Version:** v1.03
