### 2024-09-19 #7225 @cheginit
@ijbd That's awesome, welcome to the JOSS community! Sure, don't hesitate to ask questions here or reaching out via email.
