### 2024-09-28 9505-PNNL/wmpy_power#1 @ijbd
README Functionality Section Issues
![image](https://github.com/user-attachments/assets/0c09aa61-9a47-4104-9fd5-20750cf19933)

#openjournals/joss-reviews#7225
