### 2024-09-28 9505-PNNL/wmpy_power#1 @ijbd
On first read, the penstock flexibility, $f_p$, doesn't appear to be used, and an undefined subscript $f_f$ is used instead.

![image](https://github.com/user-attachments/assets/dece6045-134b-454b-8439-e5390cd38bb7)
