### 2024-09-28 9505-PNNL/wmpy_power#1 @ijbd
Similar issues on Equations 7 through 11 (although I don't see Equation 9).

![image](https://github.com/user-attachments/assets/ee105ca9-334e-455d-82c3-164ea8d50e33)
