### 2024-09-28 #7225 @ijbd
@cheginit there are two authors included in this paper who have not committed to the codebase on GitHub, I assume because this submission was repackaged. The guidelines say it is ultimately up to the submitting author to determine authorship, but would it be preferable to request a summary of contributions?
