### 2024-09-28 9505-PNNL/wmpy_power#1 @thurber
Thanks for the feedback! @danbroman can you review and update the equations as needed?
