### 2024-09-28 #7225 @thurber
hi @ijbd and all, here's a summary of contributions (not fully reflected in the git history due to transfer from a repository on a private system):

* Travis Thurber - code and paper writing
* Dan Broman - documentation, data preparation, testing, and science application
* Tian Zhou - algorithm and method creator, original Matlab code
* Nathalie Voisin - advisor, editing, ideation, and feedback
