### 2024-09-30 9505-PNNL/wmpy_power#1 @danbroman
@ijbd thanks for the review. i believe i addressed the errors you found - the Qmax formula was incomplete, and f_m should have been f_p. please let me know if you see any other issues
