### 2024-10-01 a249f97
Update np.inf for compatibility with numpy 2.0.
