### 2024-10-03 #7225 @ijbd
@cheginit, I am close to finishing my review. I had only one issue with the software, which has already been resolved. Beyond that, I especially appreciate the simple API and the documentation describing the underlying model. @thurber, thank you for your responsiveness!

Before finishing, I wanted to get either of your thoughts on test cases.

At the moment, the repository has three unit tests, two for the optimization algorithm and one for the model itself. The `ModelTest` runs through a full example of calibrating and generating data. In your opinion, is it worth expanding the test cases to individual methods? My thought is that testing individual methods provides extra documentation, which will be helpful for new and returning contributors. It would also retain visibility through big changes, reducing the burden on code reviewers. Given the niche of this codebase, I think the authors have the best sense of who is likely to make contributions and how frequently, so I'm inclined to leave it to their discretion.
