### 2024-10-30 #7225 @antviro
Hi, sorry for the delay. I have just now install the software through pip, and I have seen that as np.Inf is removed in Numpy  >2.0 (it seems that we have to use only np.inf from 2.0 and on) it fails to load the package.

I see now that it is already fixed in the repository, so I go on using the respository version, but it would be nice to "upload" it to pip system so that when people read the paper, they can install it using pip. I'll write again with more comments if needed.
