### 2024-10-30 #7225 @antviro
Just a minor comment, while following the tutorial, when plotting the calibrations with m.plot(calibrations), the default figure is a 5760x5760 with 19 axes. Figure was fine and nice, although it took me a while to see it with two not too small screens. Perhaps it would be nice for the tutorial to split into more smaller figures. Perhaps Jupyter shows them in a way it is not uncomfortable (I tested it using ipython).

I am about to finish the review. I will come back tomorrow and try to finish it. So far, I have no more comments and find the software quite useful and nice!

Best regads,
Antonio
