### 2024-10-30 #7225 @thurber
Thanks @antviro for your feedback! I will definitely release a new version to PyPI encompassing all the review feedback. I will put some thought into the size of the tutorial plot and propose some changes. It's true that jupyter limits the size but allows one to zoom in. However, it's no fun if the image takes up multiple monitors in a CLI environment!
