### 2024-11-03 #7225 @antviro
Regarding the last two items I have to check:
- documentation: perhaps it would be nice some examples, although in the notebook some examples are found, so I understand it is enough. By the way, in the default values for the start and end dates, np.Inf should be replaced with np.inf instead (coherently with the change in the code derived from the numpy change...).
- community guidelines: I understand it is enough with the github interface + the "ongoing research" part in the paper.

So, I think it is ready for publication.

Once more, sorry for the delay and thank you very much for the software release!
