### 2024-11-06 42d55b9
update np.Inf to np.inf in the readme; reduce dpi of figure
