### 2024-11-06 bfd5cae
update broman doi; update paper upload action
