+++ paper/paper.bib
@@ -1,5 +1,9 @@
 @article{Broman:2024,
-    journal = {{in prep}},
+    journal = {{Environmental Research Letters}},
+    volume = {19},
+    number = {9},
+    url = {https://doi.org/10.1088/1748-9326/ad6ceb},
+    doi = {10.1088/1748-9326/ad6ceb},
     year = {2024},
     title = {{Multi-scale impacts of climate change on hydropower for long-term water-energy planning in the contiguous United States}},
     author = {Broman, Daniel and Voisin, Nathalie and  Kao, {Shih-Chieh} and Fernandez, Alisha and Ghimire, Ganesh},
