### 2024-11-06 #7225 @thurber
Thanks all! I've reduced the size of the figure, fixed a couple doc references to `np.inf`, released a new version of wmpy-power to PyPI (v1.0.2), and updated one of the missing references in the paper.
