### 2024-11-07 #7225 @cheginit
Please address the following in the paper:

- [ ] L12: A space is missing in "wmpy-power–the"
- [ ] L23: Since you're not using "SCE" anywhere in the text, the abbreviation is not needed in "Shuffled Complex Evolution (SCE)"
- [ ] Fig 1: The legends and titles of the figure are too small and difficult to read, lease increase their size.
- [ ] L52: Since this is the first time that you're using  USDOE, please spell it out.
- [ ] L69: Please be consistent in the way that titles of References are captalized, either only the first letter, or title-case. This item uses title-case, so you can change this to only capitalizing the first letter to be consistent with the rest.
- [ ] L85: Please capitalize US.
- [ ] L92: Please capitalize System Models.

Once you addressed these issues, please create an archive on Zenodo, for example, and provide the DOI.
