### 2024-11-07 7c71031
paper updates
