+++ paper/paper.md
@@ -30,11 +30,11 @@ bibliography: paper.bib
 
 # Summary
 
-Hydropower is an important source of renewable energy in many parts of the world. The generation potential for a hydropower facility can vary greatly due to fluctuations in precipitation and snowmelt patterns impacting streamflow and reservoir storage. Human activities such as irrigation, manufacturing, and hydration can also influence water availability at nearby and downstream facilities. `wmpy-power`--the hydropower model described in this work--is process-based, leveraging explicit reservoir storage and release data to address impacts on hydropower from climate change and human adaptive behaviors to inform long-term planning and resource-adequacy considerations.
+Hydropower is an important source of renewable energy in many parts of the world. The generation potential for a hydropower facility can vary greatly due to fluctuations in precipitation and snowmelt patterns impacting streamflow and reservoir storage. Human activities such as irrigation, manufacturing, and hydration can also influence water availability at nearby and downstream facilities. `wmpy-power`, the hydropower model described in this work, is process-based, leveraging explicit reservoir storage and release data to address impacts on hydropower from climate change and human adaptive behaviors to inform long-term planning and resource-adequacy considerations.
 
 # Statement of need
 
-`wmpy-power` (Water Management Python - Hydropower) is a Python implementation of the WMP algorithm [@Zhou:2018] for hydropower simulation developed to support long-term electricity grid infrastructure planning and climate impacts studies. The model simulates hydropower production at the facility scale using a minimal set of physical characteristics for each facility, timeseries of daily streamflow and reservoir storage, and historical observations of monthly hydropower production. With this data, the model performs a two-step calibration process using the Shuffled Complex Evolution (SCE) algorithm [@Duan:1993] to optimize a set of facility and regional efficiency and bias-correction factors. Once calibrated, the model can then be used to simulate regional and facility-scale hydropower production for arbitrary timeseries of streamflow and reservoir storage. See \autoref{fig:generation} for an example of `wmpy-power` modeled generation compared to observed generation at the facility and regional scale.
+`wmpy-power` (Water Management Python - Hydropower) is a Python implementation of the WMP algorithm [@Zhou:2018] for hydropower simulation developed to support long-term electricity grid infrastructure planning and climate impacts studies. The model simulates hydropower production at the facility scale using a minimal set of physical characteristics for each facility, timeseries of daily streamflow and reservoir storage, and historical observations of monthly hydropower production. With this data, the model performs a two-step calibration process using the Shuffled Complex Evolution algorithm [@Duan:1993] to optimize a set of facility and regional efficiency and bias-correction factors. Once calibrated, the model can then be used to simulate regional and facility-scale hydropower production for arbitrary timeseries of streamflow and reservoir storage. See \autoref{fig:generation} for an example of `wmpy-power` modeled generation compared to observed generation at the facility and regional scale.
 
 ![Example model output of simulated hydropower at the regional and facility scales compared with example observations. The regional signal exhibits high fidelity despite the noise and missing data points in the certain facility signals. In this example, the calibration period was 1995-2006 and the simulation period was 2007-2010.\label{fig:generation}](figure1.png)
 
@@ -44,10 +44,10 @@ As a process-based model, `wmpy-power` utilizes time series of channel flow and 
 
 # Ongoing research
 
[…]
