+++ paper/paper.bib
@@ -25,12 +25,12 @@
     doi = {10.2172/1340431},
   	journal = {{ORNL/SR2015/357, Oak Ridge National Laboratory, Oak Ridge, TN}},
     year = {2016},
-  	title = {{The Second Assessment of the Effects of Climate Change on Federal Hydropower}},
+  	title = {The second assessment of the effects of climate change on federal hydropower},
     author = {{Kao, S.-C.} and Ashfaq, M. and Naz, B.~S. and {Martínez}, U.~R. and Rastogi, D. and Mei, R. and Jager, Y. and Samu, N.~M. and Sale, M.~J.},
 }
 
 @article{Kao:2022, 
-    title = {The Third Assessment of the Effects of Climate Change on Federal Hydropower}, 
+    title = {The third assessment of the effects of climate change on federal hydropower}, 
     author = {Kao, Shih-Chieh and Ashfaq, Moetasim and Rastogi, Deeksha and Gangrade, Sudershan and Uria Martinez, Rocio and Fernandez, Alisha and Konapala, Goutam and Voisin, Nathalie and Zhou, Tian and Xu, Wenwei and Gao, Huilin and Zhao, Bingjie and Zhao, Gang}, 
     doi = {10.2172/1887712}, 
     url = {https://www.osti.gov/biblio/1887712}, 
@@ -63,7 +63,7 @@
 }
 
 @article{Turner:2021, 
-    title = {Water storage and release policies for all large reservoirs of conterminous United States}, 
+    title = {{Water storage and release policies for all large reservoirs of conterminous United States}}, 
     journal = {Journal of Hydrology}, 
     volume = {603}, 
     pages = {126843},
@@ -86,7 +86,7 @@
 }
 
 @article{Voisin:2013,
-    title={On an improved sub-regional water resources management representation for integration into {Earth} system models.}, 
+    title={{On an improved sub-regional water resources management representation for integration into Earth System Models.}},
     author={Voisin, Nathalie and Li, Hongyi and Ward, D and Huang, Maoyi and Wigmosta, M and Leung, LR}, 
     journal={Hydrology \& Earth System Sciences}, 
     volume={17}, 
