+++ wmpy_power/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "1.0.2"
+__version__ = "1.0.3"
 from .model import Model
