### 2024-11-07 #7225 @thurber
Thanks @cheginit, I've addressed these items and released a new version of wmpy-power, v1.0.3, with the changes!

Software Version: v1.0.3
DOI: https://doi.org/10.5281/zenodo.14053013
