### 2024-11-08 #7225 @kyleniemeyer
This looks good to me, but I will wait for confirmation from @thurber about the paper.
